Draw digits in petrol station style.
Fix 'pbm2oled' so that it works correctly for images that are not a multiple of 8 pixels in width ('petrol.pbm' is 210 pixels wide). Use the new image data to generate digits 0-9 in the style of a petrol station fuel price display. Also fix a comment and eliminate a warning. TODO: add hex digits A-F in petrol station style. TODO: design other digit styles as image data until we run out of Flash ROM space.